changed avro int to long

+++ airbyte-integrations/bases/base-java-s3/src/main/java/io/airbyte/integrations/destination/s3/avro/JsonSchemaType.java
@@ -16,7 +16,7 @@ import org.apache.avro.Schema;
 public enum JsonSchemaType {
 
   STRING_V1("WellKnownTypes.json#/definitions/String", Schema.Type.STRING),
-  INTEGER_V1("WellKnownTypes.json#/definitions/Integer", Schema.Type.INT),
+  INTEGER_V1("WellKnownTypes.json#/definitions/Integer", Schema.Type.LONG),
   NUMBER_V1("WellKnownTypes.json#/definitions/Number", Schema.Type.DOUBLE),
   BOOLEAN_V1("WellKnownTypes.json#/definitions/Boolean", Schema.Type.BOOLEAN),
   BINARY_DATA_V1("WellKnownTypes.json#/definitions/BinaryData", Schema.Type.BYTES),
@@ -31,7 +31,7 @@ public enum JsonSchemaType {
   @Deprecated
   STRING_V0("string", null, Schema.Type.STRING),
   @Deprecated
-  NUMBER_INT_V0("number", "integer", Schema.Type.INT),
+  NUMBER_INT_V0("number", "integer", Schema.Type.LONG),
   @Deprecated
   NUMBER_BIGINT_V0("string", "big_integer", Schema.Type.STRING),
   @Deprecated
